mms_master-main/requirements.txt
python-dotenv>=1.0.0
streamlit==1.39.0
paho-mqtt==1.6.1
numpy>=1.26.4
pandas>=2.1.0
pymssql==2.2.10
influxdb==5.3.1
sqlalchemy>=1.4.39 